--- .envDEFAULT
@@ -1,5 +1,0 @@
-PORT=5432
-DB_HOST=localhost
-DB_USERNAME=postgres
-DB_PASSWORD=admin
-DB_NAME=talent-place-db
