+++ src/entities/pagos.ts
